+++ lib/platform/image_picker/image_picker.dart
@@ -1,0 +1,2 @@
+export 'image_picker_gateway.dart';
+export 'plugin_image_picker_gateway.dart';